backend/requirements.txt
flask
flask-cors
langchain
langchain-google-genai
langchain-core
pydantic
python-dotenv
google-generativeai
